pyproject.toml
[project]
name = "vertex-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gcsfs>=2026.2.0",
    "google-cloud-aiplatform>=1.140.0",
    "google-cloud-bigquery>=3.40.1",
    "google-cloud-bigquery-storage>=2.36.2",
    "google-cloud-logging>=3.14.0",
    "google-cloud-pipeline-components>=2.22.0",
    "google-cloud-storage>=3.9.0",
    "kfp>=2.16.0",
    "numpy>=2.4.2",
    "opentelemetry-api>=1.40.0",
    "opentelemetry-exporter-gcp-trace>=1.11.0",
    "opentelemetry-sdk>=1.40.0",
    "pandas>=3.0.1",
    "pandas-gbq>=0.34.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "scikit-learn>=1.8.0",
]
